configs: separate config parser to own file
Why:

We need config parsing in number of different utilities. Instead of
duplicating, let's create a standalone config parsing utility.

What:

- Move config parsing functions to separate file
- Call the utilities in game_io
- Separate token/id in config parsing: was after parsing in game_io code
